reorganize tests for better maintainability
