+++ .github/workflows/test.yml
@@ -29,4 +29,4 @@ jobs:
           pip install pytest pytest-asyncio
 
       - name: Run tests
-        run: pytest test.py -v
+        run: pytest tests/ -v
