+++ pyproject.toml
@@ -43,6 +43,7 @@ include = ["temporal_mcp*"]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
+testpaths = ["tests"]
 filterwarnings = [
     "ignore::pytest.PytestAssertRewriteWarning"
 ]
